Added sample data to all utilities tables, all studies tables, staff tables excluding Employees_SchoolSubjects. Added two triggers (after insert, after update) on listeners.Courses; these are to insert data into tables: staff.Employees_SchoolSubjects, listeners.Students_SchoolSubjects and listeners.StudentsRepeatedSubjects (when Students get a grade below 3.0 they will have to repeat the subject next year)  if there are no corresponding records. Added three procedures used in these triggers. Small refactoring of procedures and tables(parameters taken, data types etc.). Added(switched) up-to-date backup.

+++ SchoolDatabase/Listeners/Tables/Courses.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [listeners].[Courses]
 (
 	CourseId int IDENTITY (1,1),
-    CourseGrade float,
+    CourseGrade decimal(2,1),
     DateOfAssessment DATE,
     CourseStudentId int NOT NULL,
     CourseSchoolSubjectId varchar(7) NOT NULL,
