+++ SchoolDatabase/Utilities/Tables/ClassTypes.sql
@@ -2,6 +2,6 @@
 (
 	ClassTypeId char(1),
     ClassTypeName varchar(15) NOT NULL,
-	Weighted float,
+	Weighted decimal(2,1),
 	CONSTRAINT PK_ClassTypes PRIMARY KEY (ClassTypeId)
 )
